+++ scripts/src/commands/strapi/index.ts
@@ -1,10 +1,11 @@
 import { Command } from "commander";
-import { logProgramHelp } from "../../utils";
+import { logProgramHelp, logWarning } from "../../utils";
 import bootstrapCmd from "./bootstrap";
 import configExportCmd from "./configExport";
-import dbExportCmd from "./dbExport";
-import dbImportCmd from "./dbImport";
+import exportCmd from "./export";
+import importCmd from "./import";
 import typesGenerateCmd from "./typesGenerate";
+import chalk from "chalk";
 
 /***************************************************************************************
  * CLI
@@ -15,9 +16,17 @@ const program = new Command("strapi");
 program.description("Strapi management scripts");
 program.addCommand(bootstrapCmd);
 program.addCommand(configExportCmd);
-program.addCommand(dbExportCmd);
-program.addCommand(dbImportCmd);
+program.addCommand(exportCmd);
+program.addCommand(importCmd);
 program.addCommand(typesGenerateCmd);
+
+// error on unknown commands
+program.on("command:*", function () {
+  const cmdName = program.args.join(" ");
+
+  logProgramHelp(program, `Invalid command: ${cmdName}\n`);
+  process.exit(1);
+});
 
 export default program;
